% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnimateZFunction.R
\name{AnimateZFunction}
\alias{AnimateZFunction}
\title{Animate output for the Zadeh's principle}
\usage{
AnimateZFunction(value, FUN, knots = 10, ...)
}
\description{
`AnimateZFunction` applies the selected function to a fuzzy number using the Zadeh's principle, and then animates
the output.
}
