% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApproximationAmbValPres.R
\name{ApproximationAmbValPres}
\alias{ApproximationAmbValPres}
\title{Approximation method preserving the ambiguity and value}
\usage{
ApproximationAmbValPres(value, ...)
}
\description{
`ApproximationAmbValPres` applies the aproximation preserving ambiguity and value
}
