% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximationMehodsInside.R
\docType{data}
\name{approximationMehodsInside}
\alias{approximationMehodsInside}
\title{A vector containing names of the built-in approximation methods.}
\format{
An object of class \code{character} of length 3.
}
\usage{
approximationMehodsInside
}
\value{
This function returns a vector of strings.
}
\description{
`approximationMehodsInside` is a vector containing names of the built-in approximation methods.
}
\examples{

# check the names

approximationMehodsInside

}
\keyword{datasets}
